submission/code/prompts.py
from langchain_core.prompts import PromptTemplate

ROUTER_PROMPT_TEMPLATE = """
You are an Expert Support Triage Router.
Your task is to analyze the following user support ticket and classify it.
You must return a valid JSON object with EXACTLY the following keys and allowed values:

- "domain": "<hackerrank|claude|visa|unknown>"
- "request_type": "<product_issue|feature_request|bug|invalid>"
- "product_area": "<A brief string describing the product area, e.g., 'billing', 'access', 'assessments'>"
- "risk_score": "<Low|Medium|High>"
- "justification": "<A concise explanation of the decision & response>"

Definitions:
- High risk includes fraud, security breaches, unauthorized access, identity theft, or abusive language.
- "unknown" domain should be used if the request clearly does not relate to HackerRank, Claude, or Visa.
- If it's a generic chat or unrelated question, set domain to "unknown" and request_type to "invalid".

Ticket:
{ticket_text}

JSON Output:
"""

ROUTER_PROMPT = PromptTemplate(
    input_variables=["ticket_text"],
    template=ROUTER_PROMPT_TEMPLATE
)

GENERATOR_PROMPT_TEMPLATE = """
You are a Strict Support Agent. 
Your task is to answer the user's ticket using ONLY the provided retrieved context.

CRITICAL GUARDRAIL:
If the provided context does NOT contain the exact answer, or if the user requests an action requiring human intervention (such as billing changes, account deletion, sensitive account recovery, or managing fraud/stolen cards), you MUST abort generation and output exactly the word 'ESCALATE' and nothing else.

Context:
{context}

User Ticket:
{ticket_text}

Response:
"""

GENERATOR_PROMPT = PromptTemplate(
    input_variables=["context", "ticket_text"],
    template=GENERATOR_PROMPT_TEMPLATE
)
