submission/code/requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-huggingface
chromadb
pandas
tenacity
python-dotenv
unstructured
markdown
sentence-transformers
langchain-groq
